Fix implicit conversion

+++ projects/pandora/src/svo/sparse_voxel_dag.cpp
@@ -295,7 +295,7 @@ void compressDAGs(gsl::span<SparseVoxelDAG> svos)
 						}
 					}
 				}
-				assert(j == mutDescriptor.children.size());
+				assert(j == (int)mutDescriptor.children.size());
 
                 //}
             }
